+++ dimo/Views/MenuBar/DisplayControlCard.swift
@@ -10,13 +10,13 @@ import SwiftUI
 struct DisplayControlCard: View {
     let display: MonitorInfo
     let showPresetBar: Bool
-    let onBrightnessChange: (Double) -> Void
-    
+    let onBrightnessChange: (UInt16) -> Void
+
     @State private var sliderValue: Double = 50
     @State private var isDragging = false
-    
+
     let presets = [0, 25, 50, 75, 100]
-    
+
     var body: some View {
         VStack(alignment: .leading, spacing: 12) {
             // Display name
@@ -31,34 +31,34 @@ struct DisplayControlCard: View {
                     .foregroundStyle(.secondary)
                     .contentTransition(.numericText())
             }
-            
+
             // Brightness slider
             HStack {
                 Button("Sub", systemImage: "sun.min.fill") {
                     sliderValue = max(0, sliderValue - 1)
-                    onBrightnessChange(sliderValue)
+                    onBrightnessChange(UInt16(sliderValue))
                 }
                 .labelStyle(.iconOnly)
                 .buttonStyle(.borderless)
-                
+
                 Slider(
                     value: $sliderValue,
                     in: 0...100,
                 ) { editing in
                     isDragging = editing
                     if !editing {
-                        onBrightnessChange(sliderValue)
+                        onBrightnessChange(UInt16(sliderValue))
                     }
                 }
-                
+
                 Button("Add", systemImage: "sun.max.fill") {
                     sliderValue = min(100, sliderValue + 1)
-                    onBrightnessChange(sliderValue)
+                    onBrightnessChange(UInt16(sliderValue))
                 }
                 .labelStyle(.iconOnly)
                 .buttonStyle(.borderless)
             }
-            
+
             if showPresetBar {
                 HStack(spacing: 8) {
                     ForEach(presets, id: \.self) { preset in
@@ -66,7 +66,7 @@ struct DisplayControlCard: View {
                             withAnimation {
                                 sliderValue = Double(preset)
                             }
-                            onBrightnessChange(Double(preset))
+                            onBrightnessChange(UInt16(preset))
                         }
                         .buttonStyle(.bordered)
                         .controlSize(.small)
